fix: JKRExpHeap 64-bit pointer truncation — u32 start/endAddr/nextAddr → uintptr_t, add profile list init

+++ src/JSystem/JKernel/JKRHeap.cpp
@@ -295,7 +295,7 @@ u8 JKRHeap::getCurrentGroupId() {
 }
 
 u32 JKRHeap::getMaxAllocatableSize(int alignment) {
-    u32 maxFreeBlock = (uintptr_t)getMaxFreeBlock();
+    uintptr_t maxFreeBlock = (uintptr_t)getMaxFreeBlock();
     u32 ptrOffset = (alignment - 1) & alignment - (maxFreeBlock & 0xf);
     return ~(alignment - 1) & (getFreeSize() - ptrOffset);
 }
